requirements.txt
Django>=1.5
django-crispy-forms>=1.4.0
django-registration>=1.0
